+++ src/grid/columns.js
@@ -167,6 +167,16 @@ export default {
 
     },
 
+    // Sync both width and height for all column headers after column width changes
+    updateAllColumnHeadersSize: function() {
+        const columns = this.viewColumns;
+        for (let i = 0, l = columns.length; i < l; i++) {
+            const col = columns[i];
+            this.updateColumnHeaderWidth(col);
+            this.updateColumnHeaderHeight(col, true);
+        }
+    },
+
     // update height depends width and display state
     // because width could be 0, then column is hidden
     updateColumnHeaderHeight: function(columnItem, force) {